requirements.txt
pandas
mlflow
notebook
numpy
scikit-learn
matplotlib
python-box
pyYAML
tqdm
ensure
joblib
types-pyYAML
Flask
Flask-Cors
pylint
pytest
pytest-cov
black
box==3.2.2
isort
pyarrow
invoke
jupyter
jupyterlab
ipykernel
ipython
-e .